add some ideas for future improvements
I was going to mention things like transaction relay and peer discovery
(and peer eviction), but on further thought, I don't think those belong
in this simulator, because they don't have the time element that this
simulator is really trying to explore. Transactions will eventually be
relayed; peers will eventually be discovered.